Re-writing file
Separating the macros and the main dialogue was difficult to implement the pickle functions
